Shift down LoadNetwork to auto_exec_network like hetero did
